maira_2/image_text_to_text/pytorch/requirements.txt
pillow
protobuf
sentencepiece
transformers>=4.48.0,<4.52
